# docs/swagger.yaml
basePath: /
...
host: kawai-be.vercel.app
info:
  contact:
    name: Dwi Puspa Firdaus
  description: Dokumentasi API Backend untuk Kawai Assistant (Bot WhatsApp & Dashboard).
  title: Kawai Assistant API
  version: 1.0.0
paths:
...
  /api/auth/google/callback:
    get:
      description: Menerima kode otorisasi Google dan menukar dengan Token
      parameters:
      - description: Authorization Code
        in: query
        name: code
        required: true
        type: string
      - description: User Phone Number
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: Halaman Sukses HTML
          schema:
            type: string
      summary: Callback Google Auth
      tags:
      - Auth
  /api/auth/google/login:
    get:
      description: Redirect ke halaman Login Google (OAuth2)
      parameters:
      - description: Nomor HP User (State)
        in: query
        name: phone
        required: true
        type: string
      responses:
        "307":
          description: Redirect ke Google
      summary: Login Google
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"